requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
bandit==1.8.5
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitarray==3.4.3
black==25.1.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cloudpathlib==0.21.1
cmudict==1.0.32
coloredlogs==15.0.1
confection==0.1.5
courlan==1.3.2
coverage==7.9.1
cryptography==45.0.4
cssselect==1.3.0
cymem==2.0.11
Cython==3.0.11
datasets==3.6.0
datasketch==1.6.5
dateparser==1.2.1
detoxify==0.5.2
dill==0.3.8
distlib==0.3.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
execnet==2.1.1
faiss-cpu==1.11.0
fastapi==0.115.13
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hiredis==3.2.1
hpack==4.1.0
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
hypothesis==6.135.15
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
jusText==3.0.2
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
language_tool_python==2.9.4
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.1
multiprocess==0.70.16
murmurhash==1.0.13
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pbr==6.1.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polars==1.31.0
pre_commit==4.2.0
preshed==3.0.10
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybloom_live==4.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-benchmark==5.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
# Editable install with no version control (quarrycore==0.1.0)
-e /home/admin/projects/QuarryCore
RapidFuzz==3.13.0
readability-lxml==0.8.4.1
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
selectolax==0.3.30
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
stevedore==5.4.1
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
